Correct 0x08 CP_ClickWindow for protocol 1.13.2

+++ mcp_packet.h
@@ -395,7 +395,7 @@ typedef struct {
     int16_t  sid;
     uint8_t  button;
     uint16_t aid;
-    uint8_t  mode;
+    uint32_t mode;
     slot_t   slot;
 } CP_ClickWindow_pkt;
